Made basic modules

+++ modular-io.c
@@ -2,19 +2,19 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int *read_array(int *size) {
-    scanf("%d", size);
+int *read_array(size_t *size) {
+    scanf("%lu", size);
 
     int *array = malloc(*size * sizeof(int));
-    for (int i = 0; i < *size; ++i) {
+    for (size_t i = 0; i < *size; ++i) {
         scanf("%d", &array[i]);
     }
 
     return array;
 }
 
-void print_array(int *array, int size) {
-    for (int i = 0; i < size; ++i) {
+void print_array(int *array, size_t size) {
+    for (size_t i = 0; i < size; ++i) {
         printf("%d ", array[i]);
     }
     printf("\n");
