+++ modular-io.h
@@ -1,7 +1,9 @@
 #ifndef TASK1_MODULAR_IO_H
 #define TASK1_MODULAR_IO_H
 
-int *read_array(int *size);
-void print_array(int *array, int size);
+#include <stddef.h>
+
+int *read_array(size_t *size);
+void print_array(int *array, size_t size);
 
 #endif//TASK1_MODULAR_IO_H
