+++ modular-make.h
@@ -1,6 +1,8 @@
 #ifndef TASK1_MODULAR_MAKE_H
 #define TASK1_MODULAR_MAKE_H
 
-int *make_array(const int *src_array, int src_size, int *new_size);
+#include <stddef.h>
+
+int *make_array(const int *src_array, size_t src_size, size_t *new_size);
 
 #endif//TASK1_MODULAR_MAKE_H
